pyproject.toml
[project]
name = "hackathoh-versailles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.183.0",
    "fastapi>=0.118.0",
    "llama-index>=0.14.3",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-mistralai>=0.8.1",
    "llama-index-vector-stores-weaviate>=0.8.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=3.0.0",
    "weaviate-client>=4.0.0",
    "mistralai>=1.0.0",
    "uvicorn>=0.37.0",
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",
    "tqdm>=4.66.0",
]
